+++ CHANGELOG.md
@@ -6,6 +6,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ### Added
 
+- Log parallel optimization.
 - Plot Gaussian process from hyperparameter Bayesian optimization.
 - Plot cluster losses and population histogram using matplotlib.
 - Option to use a sequential reduction optimizer for Bayesian optimization.